+++ jwst/resample/resample_spec_step.py
@@ -8,7 +8,6 @@ from . import resample_spec, ResampleStep
 from ..exp_to_source import multislit_to_container
 from ..assign_wcs.util import update_s_region_spectral
 
-
 # Force use of all DQ flagged data except for DO_NOT_USE and NON_SCIENCE
 GOOD_BITS = '~DO_NOT_USE+NON_SCIENCE'
 
@@ -30,12 +29,23 @@ class ResampleSpecStep(ResampleStep):
         self.wht_type = self.weight_type
         input_new = datamodels.open(input)
 
+        # Check if input_new is a MultiSlitModel
+        model_is_msm = isinstance(input_new, MultiSlitModel)
+
+        #  If input is a 3D rateints MultiSlitModel (unsupported) skip the step
+        if model_is_msm and len((input_new[0]).shape) == 3:
+            self.log.warning('Resample spec step will be skipped')
+            input_new.meta.cal_step.resample_spec = 'SKIPPED'
+
+            return input_new
+
         # Convert ImageModel to SlitModel (needed for MIRI LRS)
         if isinstance(input_new, ImageModel):
             input_new = datamodels.SlitModel(input_new)
 
         if isinstance(input_new, ModelContainer):
             input_models = input_new
+
             try:
                 output = input_models.meta.asn_table.products[0].name
             except AttributeError:
@@ -83,6 +93,7 @@ class ResampleSpecStep(ResampleStep):
 
         # Call resampling
         self.drizpars = kwargs
+
         if isinstance(input_models[0], MultiSlitModel):
             result = self._process_multislit(input_models)